+++ src/player/player_context.cc
@@ -12,17 +12,17 @@
 PlayerContext PlayerContext::default_player_context_{led::kUndefinedLedPin, 0};
 
 PlayerContext::PlayerContext(
-    std::uint32_t led_pin,
-    std::uint32_t remaining_seconds) : led_pin_(led_pin), remaining_seconds_(remaining_seconds) {}
+    std::uint16_t led_pin,
+    std::uint16_t remaining_seconds) : led_pin_(led_pin), remaining_seconds_(remaining_seconds) {}
 
 std::int16_t PlayerContext::led_pin() {
   return led_pin_;
 }
 
-std::uint32_t PlayerContext::remaining_seconds() {
+std::uint16_t PlayerContext::remaining_seconds() {
   return remaining_seconds_;
 }
 
-void PlayerContext::remaining_seconds(std::uint32_t remaining_seconds) {
+void PlayerContext::remaining_seconds(std::uint16_t remaining_seconds) {
   remaining_seconds_ = remaining_seconds;
 }
